Move checksum utils into their own file
