+++ include/tins/CMakeLists.txt
@@ -7,3 +7,4 @@ INSTALL(
 ADD_SUBDIRECTORY(dot11)
 ADD_SUBDIRECTORY(tcp_ip)
 ADD_SUBDIRECTORY(detail)
+ADD_SUBDIRECTORY(utils)
